+++ test/end2end/shared.dart
@@ -2,7 +2,6 @@ class E2EData {
   final String name;
   final String version;
   final Map<String, dynamic> data;
-  final Map<String, String> descriptions;
 
-  E2EData(this.name, this.version, this.data, this.descriptions);
+  E2EData(this.name, this.version, this.data);
 }