+++ test/platform_test.dart
@@ -6,10 +6,10 @@ import 'pubspec_test.dart';
 void main() {
   group('Platform', () {
     test('no libraries', () {
-      var p = classifyPlatform([]);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform([]);
       expect(p.worksEverywhere, isTrue);
+      expect(p.restrictedTo, isNull);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isTrue);
       expect(p.worksOnFlutter, isTrue);
@@ -17,10 +17,10 @@ void main() {
     });
 
     test('unknown library', () {
-      var p = classifyPlatform(['package:_unknown/_unknown.dart']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['package:_unknown/_unknown.dart']);
       expect(p.worksEverywhere, isTrue);
+      expect(p.restrictedTo, isNull);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isTrue);
       expect(p.worksOnFlutter, isTrue);
@@ -28,30 +28,30 @@ void main() {
     });
 
     test('dart:io', () {
-      var p = classifyPlatform(['dart:io']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:io']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['flutter', 'server']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isTrue);
       expect(p.worksOnFlutter, isTrue);
-      expect(p.description, 'flutter, server');
+      expect(p.description, 'flutter,server');
     });
 
     test('dart:html', () {
-      var p = classifyPlatform(['dart:html']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:html']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['web']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
       expect(p.description, 'web');
 
-      p = classifyPlatform(['dart:svg']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      p = classifyLibPlatform(['dart:svg']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['web']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
@@ -59,10 +59,10 @@ void main() {
     });
 
     test('dart:ui', () {
-      var p = classifyPlatform(['dart:ui']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:ui']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['flutter']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isTrue);
@@ -70,21 +70,21 @@ void main() {
     });
 
     test('dart:mirrors', () {
-      var p = classifyPlatform(['dart:mirrors']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:mirrors']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['server', 'web']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isTrue);
       expect(p.worksOnFlutter, isFalse);
-      expect(p.description, 'server, web');
+      expect(p.description, 'server,web');
     });
 
     test('http package: both html and io', () {
-      var p = classifyPlatform(['dart:html', 'dart:io']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:html', 'dart:io']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['web']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isTrue);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
@@ -92,98 +92,89 @@ void main() {
     });
 
     test('detect native', () {
-      var p = classifyPlatform(['dart:io', 'dart-ext:some-extension']);
-      expect(p.hasConflict, isFalse);
-      expect(p.worksAnywhere, isTrue);
+      var p = classifyLibPlatform(['dart:io', 'dart-ext:some-extension']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, ['server']);
+      expect(p.hasConflict, isFalse);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isTrue);
       expect(p.worksOnFlutter, isFalse);
-      expect(p.uses, [PlatformFlags.dartExtension, 'dart:io']);
       expect(p.description, 'server');
     });
   });
 
   group('Conflicting Platform', () {
     test('dart:html + dart:ui', () {
-      var p = classifyPlatform(['dart:html', 'dart:ui']);
-      expect(p.hasConflict, isTrue);
-      expect(p.worksAnywhere, isFalse);
+      var p = classifyLibPlatform(['dart:html', 'dart:ui']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, isNull);
+      expect(p.hasConflict, isTrue);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
-      expect(p.description, 'conflict');
+      expect(p.description, 'undefined');
     });
 
     test('dart:mirrors + dart:ui', () {
-      var p = classifyPlatform(['dart:mirrors', 'dart:ui']);
-      expect(p.hasConflict, isTrue);
-      expect(p.worksAnywhere, isFalse);
+      var p = classifyLibPlatform(['dart:mirrors', 'dart:ui']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, isNull);
+      expect(p.hasConflict, isTrue);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
-      expect(p.description, 'conflict');
+      expect(p.description, 'undefined');
     });
 
     test('native + dart:ui', () {
-      var p = classifyPlatform(['dart:ui', 'dart-ext:some-extension']);
-      expect(p.hasConflict, isTrue);
-      expect(p.worksAnywhere, isFalse);
+      var p = classifyLibPlatform(['dart:ui', 'dart-ext:some-extension']);
       expect(p.worksEverywhere, isFalse);
+      expect(p.restrictedTo, isNull);
+      expect(p.hasConflict, isTrue);
       expect(p.worksOnWeb, isFalse);
       expect(p.worksOnServer, isFalse);
       expect(p.worksOnFlutter, isFalse);
-      expect(p.uses, [PlatformFlags.dartExtension, 'dart:ui']);
-      expect(p.description, 'conflict');
+      expect(p.description, 'undefined');
     });
   });
 
-  group('PlatformSummary', () {
+  group('PkgPlatform', () {
     test('handles multiple libraries', () {
-      var sum = classifyPlatforms(emptyPubspec, {
+      var sum = classifyPkgPlatform(emptyPubspec, {
         'package:_example/a.dart': ['dart:html'],
         'package:_example/b.dart': ['dart:io'],
       });
-      expect(sum.hasConflict, isFalse);
-      expect(sum.pubspec.isFlutter, isFalse);
-      expect(sum.libraries.length, 2);
-      expect(sum.pubspec.description, 'undefined');
-
-      var pa = sum.libraries['package:_example/a.dart'];
-      expect(pa.worksOnWeb, isTrue);
-      expect(pa.worksOnServer, isFalse);
-      expect(pa.description, 'web');
-
-      var pb = sum.libraries['package:_example/b.dart'];
-      expect(pb.worksOnWeb, isFalse);
-      expect(pb.worksOnServer, isTrue);
-      expect(pb.description, 'flutter, server');
+      expect(sum.worksEverywhere, isFalse);
+      expect(sum.restrictedTo, ['flutter', 'server', 'web']);
+      expect(sum.descriptionAndReason, 'flutter,server,web: all of the above');
     });
 
     test('detects flutter in pubspec', () {
-      var sum = classifyPlatforms(flutterPluginPubspec, {});
-      expect(sum.hasConflict, isFalse);
-      expect(sum.pubspec.isFlutter, isTrue);
-      expect(sum.pubspec.description, 'flutter');
+      var sum = classifyPkgPlatform(flutterPluginPubspec, {});
+      expect(sum.worksEverywhere, isFalse);
+      expect(sum.restrictedTo, ['flutter']);
+      expect(sum.descriptionAndReason,
+          'flutter: pubspec reference with no conflicts');
     });
 
     test('detects flutter in dependencies', () {
-      var sum = classifyPlatforms(flutterSdkPubspec, {});
-      expect(sum.hasConflict, isFalse);
-      expect(sum.pubspec.isFlutter, isTrue);
-      expect(sum.pubspec.description, 'flutter');
+      var sum = classifyPkgPlatform(flutterSdkPubspec, {});
+      // TODO: fix logic, detect SDK
+      expect(sum.worksEverywhere, isFalse);
+      expect(sum.restrictedTo, isNull);
+      expect(sum.descriptionAndReason, 'undefined: no libraries!');
     });
   });
 
-  group('Conflicting PlatformSummary', () {
+  group('Conflicting PkgPlatform', () {
     test('Flutter package with mirrors', () {
-      var sum = classifyPlatforms(flutterPluginPubspec, {
+      var sum = classifyPkgPlatform(flutterPluginPubspec, {
         'package:_example/lib.dart': ['dart:mirrors'],
       });
-      expect(sum.pubspec.description, 'flutter');
-      expect(sum.hasConflict, isTrue);
+      expect(sum.worksEverywhere, isFalse);
+      expect(sum.restrictedTo, isNull);
+      expect(sum.descriptionAndReason,
+          'undefined: flutter reference with library conflicts');
     });
   });
 }